Actually respect --no-delete !

+++ hsync/dest_impl.py
@@ -195,7 +195,7 @@ def _fetch_remote_impl(needed, not_needed, dst_hashlist, abs_hashfile, opt):
     (fetch_added, fetch_err_count) = fetch_needed(needed, opt.source_url, opt)
 
     # Don't delete things if we had transfer problems. It's safer.
-    if fetch_err_count == 0:
+    if not opt.no_delete and fetch_err_count == 0:
         delete_status = delete_not_needed(not_needed, opt.dest_dir, opt)
 
     if (fetch_err_count > 0 or not delete_status):
